apps/projects/modules: move module/timeline logic to mixin
