+++ meinberlin/apps/projects/templates/meinberlin_projects/project_detail.html
@@ -154,7 +154,6 @@
     role="tabpanel"
     aria-labelledby="tab-project-{{ project.pk }}-participation"
     aria-expanded="true">
-
         <div class="l-wrapper">
             <div class="timeline-carousel__wrapper">
             {% if view.display_timeline %}
